Use command pattern for rotation action
