+++ Assets/__Scripts/Editor/Input/BeatmapRotationInputController.cs
@@ -11,7 +11,6 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
 {
     [SerializeField] private AudioTimeSyncController atsc;
     [SerializeField] private ScrollPrecisionController scrollPrecisionController;
-    [SerializeField] private TracksManager tracksManager;
     [SerializeField] private LaneRotationProvider laneRotationProvider;
 
     public void OnRotateClockwise(InputAction.CallbackContext context)
@@ -19,17 +18,18 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
         if (CustomStandaloneInputModule.IsPointerOverGameObject<GraphicRaycaster>(0, true)) return;
         if (!context.performed || !EditContext.EditingMode.HasFlag(editMode) || !RaycastFirstObject(out var con))
             return;
+        var prec = scrollPrecisionController.GetCurrentRotationPrecision();
         switch (con)
         {
             case RotationEventContainer:
-                RotateObject(con, true);
+                RotationCommand.RotateObject(con.ObjectData, true, prec);
                 break;
             case NoteContainer:
             case ObstacleContainer:
             case ArcContainer:
             case ChainContainer:
                 if (BeatSaberSongContainer.Instance.Map.MajorVersion != 4) return;
-                RotateObject(con, true);
+                RotationCommand.RotateObject(con.ObjectData, true, prec);
                 break;
         }
     }
@@ -39,17 +39,18 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
         if (CustomStandaloneInputModule.IsPointerOverGameObject<GraphicRaycaster>(0, true)) return;
         if (!context.performed || !EditContext.EditingMode.HasFlag(editMode) || !RaycastFirstObject(out var con))
             return;
+        var prec = scrollPrecisionController.GetCurrentRotationPrecision();
         switch (con)
         {
             case RotationEventContainer:
-                RotateObject(con, false);
+                RotationCommand.RotateObject(con.ObjectData, false, prec);
                 break;
             case NoteContainer:
             case ObstacleContainer:
             case ArcContainer:
             case ChainContainer:
                 if (BeatSaberSongContainer.Instance.Map.MajorVersion != 4) return;
-                RotateObject(con, false);
+                RotationCommand.RotateObject(con.ObjectData, false, prec);
                 break;
         }
     }
@@ -80,39 +81,6 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
                 Mathf.Round((laneRotationProvider.EditRotation - prec) * 1_000f) / 1_000f));
     }
 
-    private void RotateObject(ObjectContainer c, bool clockwise)
-    {
-        var originalObject = c.ObjectData;
-        var newObject = BeatmapFactory.Clone(originalObject);
-
-        var modifier = clockwise ? 1 : -1;
-        switch (newObject)
-        {
-            case BaseRotationEvent evt:
-                {
-                    var prec = scrollPrecisionController.GetCurrentRotationPrecision();
-                    evt.Rotation = Mathf.Round((evt.Rotation + (modifier * prec)) * 1_000f) / 1_000f;
-                    break;
-                }
-            case BaseGrid grid:
-                {
-                    var prec = scrollPrecisionController.GetCurrentRotationPrecision();
-                    grid.Rotation += Mathf.RoundToInt(modifier * prec);
-                    break;
-                }
-        }
-
-        if (newObject.CompareTo(originalObject) == 0) return;
-
-        BeatmapActionContainer.AddAction(
-            new BeatmapObjectUpdatedAction(
-                newObject,
-                originalObject,
-                mergeType: ActionMergeType.ModifyRotationValue),
-            perform: true);
-        tracksManager.RefreshTracks();
-    }
-
     public void OnInvert(InputAction.CallbackContext context)
     {
         if (CustomStandaloneInputModule.IsPointerOverGameObject<GraphicRaycaster>(0, true)) return;
@@ -122,7 +90,7 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
             || con is not RotationEventContainer e)
             return;
 
-        Invert(e);
+        RotationCommand.Invert(e.EventData);
     }
 
     public void OnModifyHover(InputAction.CallbackContext context)
@@ -135,41 +103,7 @@ public class BeatmapRotationInputController : BeatmapInputController<ObjectConta
             return;
 
         var modifier = context.GetScrollDirection(Settings.Instance.InvertScrollEventValue);
-        ModifyHover(e, modifier);
-    }
-
-    public void Invert(RotationEventContainer e)
-    {
-        var originalObject = e.EventData;
-        var newObject = BeatmapFactory.Clone(originalObject);
-
-        newObject.Rotation *= -1;
-
-        BeatmapActionContainer.AddAction(
-            new BeatmapObjectUpdatedAction(
-                newObject,
-                originalObject,
-                mergeType: ActionMergeType.ModifyRotationValue),
-            perform: true);
-        tracksManager.RefreshTracks();
-    }
-
-    private void ModifyHover(RotationEventContainer e, int modifier)
-    {
-        var originalObject = e.EventData;
-        var newObject = BeatmapFactory.Clone(originalObject);
-
         var prec = scrollPrecisionController.GetCurrentRotationPrecision();
-        newObject.Rotation = Mathf.Round((newObject.Rotation + (modifier * prec)) * 1_000f) / 1_000f;
-
-        if (newObject.CompareTo(originalObject) == 0) return;
-
-        BeatmapActionContainer.AddAction(
-            new BeatmapObjectUpdatedAction(
-                newObject,
-                originalObject,
-                mergeType: ActionMergeType.ModifyRotationValue),
-            perform: true);
-        tracksManager.RefreshTracks();
+        RotationCommand.ModifyHover(e.EventData, modifier, prec);
     }
 }
